+++ database/migrations/2023_02_17_091650_create_water_histories_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('water_histories', function (Blueprint $table) {
             $table->id();
             $table->unsignedBigInteger('water_way_id');
-            $table->bigInteger('volume');
+            $table->float('volume');
             $table->timestamps();
         });
     }
